# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TransportExperience API",
    "description": "\nBackend del MVP de TransportExperience: alquiler, venta y domicilio de ciclas, patines y\nmonopatines eléctricos.\n\n**Autenticación.** `POST /autenticacion/token` con un formulario `username` (el correo) y\n`password`. Envía el `access_token` recibido en el encabezado `Authorization: Bearer <token>`.\nEn esta página, usa el botón **Authorize**.\n\n**Errores.** Todas las respuestas de error tienen la forma\n`{\"error\": {\"codigo\", \"mensaje\", \"accion\", \"detalles\"}}`. Decide por `codigo`; muestra\n`mensaje` y `accion` al usuario.\n\n**Listas.** Siempre paginadas: parámetros `pagina` (desde 1) y `tamano` (máx. 100).\n",
    "version": "0.1.0"
  },
  "paths": {
    "/autenticacion/token": {
      "post": {
        "tags": [
          "autenticación"
        ],
        "summary": "Iniciar sesión",
        "description": "Recibe un formulario (``application/x-www-form-urlencoded``) con ``username`` = el\ncorreo y ``password``. Devuelve el token para enviar en ``Authorization: Bearer``.",
        "operationId": "iniciar_sesion_autenticacion_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_iniciar_sesion_autenticacion_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenAcceso"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado o sesión expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
    "/salud": {
      "get": {
        "tags": [
          "sistema"
        ],
        "summary": "Chequeo de salud",
        "description": "Indica si la API está viva y si alcanza la base de datos.",
        "operationId": "salud_salud_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoSalud"
                }
              }
            }
          },
          "503": {
            "description": "La base de datos no responde",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoSalud"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_iniciar_sesion_autenticacion_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_iniciar_sesion_autenticacion_token_post"
      },
      "CuerpoError": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "accion": {
            "type": "string",
            "title": "Accion"
          },
          "detalles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetalleCampo"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalles"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "mensaje",
          "accion"
        ],
        "title": "CuerpoError"
      },
      "DetalleCampo": {
        "properties": {
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          }
        },
        "type": "object",
        "required": [
          "campo",
          "mensaje"
        ],
        "title": "DetalleCampo"
      },
      "ErrorRespuesta": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/CuerpoError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorRespuesta",
        "description": "Forma de todas las respuestas de error de la API."
      },
      "EstadoSalud": {
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "ok",
              "degradado"
            ],
            "title": "Estado"
          },
          "base_datos": {
            "type": "string",
            "enum": [
              "ok",
              "sin_conexion"
            ],
            "title": "Base Datos"
          }
        },
        "type": "object",
        "required": [
          "estado",
          "base_datos"
        ],
        "title": "EstadoSalud",
        "description": "Respuesta del chequeo de salud."
      },
...
      "Rol": {
        "type": "string",
        "enum": [
          "cliente",
          "administrador",
          "operador",
          "domiciliario"
        ],
        "title": "Rol",
        "description": "Los cuatro roles del SRS. HU-12 nombra tres; el domiciliario lo exigen HU-06 y HU-07."
      },
      "TokenAcceso": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expira_en_segundos": {
            "type": "integer",
            "title": "Expira En Segundos",
            "description": "Tras este tiempo hay que iniciar sesión de nuevo."
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioRespuesta"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expira_en_segundos",
          "usuario"
        ],
        "title": "TokenAcceso",
        "description": "Respuesta del inicio de sesión (formato OAuth2, más los datos del usuario)."
      },
...
      "UsuarioRespuesta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre_completo": {
            "type": "string",
            "title": "Nombre Completo"
          },
          "correo": {
            "type": "string",
            "title": "Correo"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "rol": {
            "$ref": "#/components/schemas/Rol"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          },
          "creado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creado En"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre_completo",
          "correo",
          "telefono",
          "rol",
          "activo",
          "creado_en"
        ],
        "title": "UsuarioRespuesta",
        "description": "Usuario tal como lo ve la API. Nunca incluye el hash de la contraseña."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/autenticacion/token"
          }
        }
      }
    }
  }
}